+++ edit.php
@@ -1,0 +1,54 @@
+<?php
+// This file is part of Moodle - http://moodle.org/
+//
+// Moodle is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// Moodle is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
+
+/**
+ * Content authoring page for mod_elang.
+ *
+ * Ensures the activity has a draft version to edit (branching a copy from the
+ * published version when there is one) and hands off to the mod_elang/editor
+ * AMD module, which loads the draft through the external API and drives editing,
+ * subtitle import, saving and publishing. This is the editor foundation; the
+ * timeline interactions are layered on in later slices.
+ *
+ * @package    mod_elang
+ * @copyright  2026 Ralf Erlebach
+ * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
+ */
+
+require(__DIR__ . '/../../config.php');
+
+$id = required_param('id', PARAM_INT);
+
+[$course, $cm] = get_course_and_cm_from_cmid($id, 'elang');
+$elang = $DB->get_record('elang', ['id' => $cm->instance], '*', MUST_EXIST);
+
+require_login($course, true, $cm);
+$context = context_module::instance($cm->id);
+require_capability('mod/elang:manage', $context);
+
+$PAGE->set_url('/mod/elang/edit.php', ['id' => $cm->id]);
+$PAGE->set_title(format_string($elang->name));
+$PAGE->set_heading(format_string($course->fullname));
+$PAGE->set_context($context);
+$PAGE->set_activity_record($elang);
+
+$draft = (new \mod_elang\local\domain\version_manager())->get_or_create_draft((int) $elang->id);
+
+$PAGE->requires->js_call_amd('mod_elang/editor', 'init', [(int) $cm->id, (int) $draft->id]);
+
+echo $OUTPUT->header();
+echo $OUTPUT->render_from_template('mod_elang/editor', []);
+echo $OUTPUT->footer();